Add security

+++ products/views.py
@@ -4,6 +4,7 @@ from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
 from .forms import ProductForm
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -71,8 +72,12 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
+@login_required
 def add_product(request):
     """Add a product to the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'You are not authorised to do this')
+        return redirect(reverse('home'))
     if request.method == 'POST':  # post handler
         form = ProductForm(request.POST, request.FILES)  # new instance of product form from request.Post and request.files (images)
         if form.is_valid():
@@ -91,8 +96,12 @@ def add_product(request):
 
     return render(request, template, context)
 
+@login_required
 def edit_product(request, product_id):
     """edit product"""
+    if not request.user.is_superuser:
+        messages.error(request, 'You are not authorised to do this')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == 'POST':
@@ -115,8 +124,12 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+@login_required
 def delete_product(request, product_id):
     """delete product"""
+    if not request.user.is_superuser:
+        messages.error(request, 'You are not authorised to do this')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted')  # product.name raises error
